requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
psycopg2-binary==2.9.10
